refactor: move CryptoService to domain
