+++ ProtoCredentials/OpenID4VC-Prototype/OpenID4VC-Prototype.csproj
@@ -25,7 +25,6 @@
 
 	<ItemGroup>
 	  <Folder Include="Application\Services\" />
-	  <Folder Include="Domain\Services\" />
 	  <Folder Include="Infrastructure\Configurations\" />
 	  <Folder Include="Infrastructure\Interfaces\" />
 	</ItemGroup>
