+++ ProtoCredentials/OpenID4VC-Prototype/Program.cs
@@ -2,6 +2,7 @@
 using Microsoft.Extensions.Hosting;
 using OpenID4VC_Prototype.Application.Interfaces;
 using OpenID4VC_Prototype.Domain.Interfaces;
+using OpenID4VC_Prototype.Domain.Services;
 
 var builder = Host.CreateDefaultBuilder(args)
     .UseSerilog((hostContext, loggerConfiguration) =>
